refactor: 요청 헤더 포장 방식 수정
