--- tomcat/src/main/java/org/apache/coyote/http11/general/HttpHeader.java
@@ -1,5 +1,0 @@
-package org.apache.coyote.http11.general;
-
-public record HttpHeader(String key, String value) {
-
-}
