+++ tomcat/src/main/java/org/apache/coyote/http11/handler/controllerResponse/ControllerResponse.java
@@ -1,12 +1,13 @@
 package org.apache.coyote.http11.handler.controllerResponse;
 
-import java.util.List;
-import org.apache.coyote.http11.general.HttpHeader;
+import org.apache.coyote.http11.general.HttpHeaders;
 import org.apache.coyote.http11.httpResponse.HttpStatus;
 
 public interface ControllerResponse {
 
     HttpStatus status();
-    List<HttpHeader> headers();
+    HttpHeaders headers();
     String content();
+
+    void addHeader(String key, String value);
 }
