+++ tomcat/src/main/java/org/apache/coyote/http11/handler/controllerResponse/StaticFileResponse.java
@@ -1,13 +1,17 @@
 package org.apache.coyote.http11.handler.controllerResponse;
 
-import java.util.ArrayList;
-import java.util.List;
-import org.apache.coyote.http11.general.HttpHeader;
+import java.util.HashMap;
+import org.apache.coyote.http11.general.HttpHeaders;
 import org.apache.coyote.http11.httpResponse.HttpStatus;
 
-public record StaticFileResponse(HttpStatus status, List<HttpHeader> headers, String content) implements ControllerResponse {
+public record StaticFileResponse(HttpStatus status, HttpHeaders headers, String content) implements ControllerResponse {
 
     public StaticFileResponse(HttpStatus status, String content) {
-        this(status, new ArrayList<>(), content);
+        this(status, new HttpHeaders(new HashMap<>()), content);
+    }
+
+    @Override
+    public void addHeader(String key, String value) {
+        headers.add(key, value);
     }
 }
